Added verbal output of the simulator system clock
The simulator FSX/P3D has a scenario clock that allows a pilot to set the time and date of the flight. By default, it it is set to the user's desktop time. Unfortunately, it is easy to ignore the scenario clock and fly. It is important for flight tracking apps such as smart cars or acars to have the correct time and date of the flight. This update adds the verbal readouts for the system clock. NOTE: the readout is with ] F10, and is only the current local time. EX: If you fly over Michigan, the local time is announced when pressing ] F10. Zulu time will come later.