PMDG 777: autobrake
Added the PMDG 777 autobrake. The autobrake is a series of radio buttons that control the autobrake on the PMDG 777. Pressing TAB or SHIFT+TAB to the autobrake group will allow a pilot to use the arrow keys to change the current autobrake state. TFM will also announce the current autobrake state when it automatically changes from an external source such as First officer pro, it is directly changed in the simulator by the pilot, or TFM changes it. The following hotkeys are also available from anywhere in the speed box.

ALT+T = autobrake RTO.
ALT+O = autobrake off.
ALT+D = autobrake disarm.
ALT+1 = autobrake 1 (minimum).
ALT+2 = autobrake 2 (medium).
ALT+3 = autobrake 3 (maximum or auto).