ILS: Made aircraft primary focus
In the previous release, the ILS had primary focus in the world. For example, announcements were as follows: "1000 above." Instead of meaning the aircraft is 1000 feet above the glide slope, it meant the glide slope was 1000 feet above the aircraft. This fixes the problem by making the aircraft the focus of concern. Now "1000 above" means the aircraft is 1000 feet above the glide slope.