PMDG 777: Navigation aids for MCP readouts
Added the functions needed to announce the navigation aid currently in use when the pilot uses left bracket ([), then the letters a for altitude and h for heading. Sample announcements will look like the below.

FMC speed when speed intervene is turned off.
MCP speed 310 for a manually entered indicated airspeed.
MCP speed .79 for a manually entered mach speed.
MCP altitude 2000 VNav if VNav is in use
MCP altitude 6000 feet level change if level change is turned on.

MCP heading 170 LNav if LNav is turned on
MCP heading 200 hold if heading select/hold is on.
MCP heading 022 combined if both LNav and heading hold are both on.