TFM: Introduce database storage support
Introduces database storage capabilities. We start by focusing on speech history in this build. The new features include the following.
Brings 2 new settings to the output category. 'Save output to a database' and 'Include timestamps in the speech history'. When checked, 'save output to database' will make use of a TFM generated database to store different information. When not checked, TFM will revert to its current behavior. The speech history size control in the output category in settings will not be available when using database support. When using database support, a new feature, 'Include timestamps in speech history' will become available. When checked, timestamps are included in the speech history found by pressing right bracket (]) then CTRL+S while TFM is running. When not checked, the speech history behaves the same as before with the exception that history is kept forever, and across restarts.
TFM automatically generates the database and everything it requires to use the new database features. The current location for the database is the user's documents folder under Talking flight monitor/data/TFM.db. Manually deleting this file or folder structure will force TFM to regenerate the database on next restart. To stop regenerating the database after it is deleted, uncheck the 'Save output to a database' box in output category in TFM settings before deleting it and restarting TFM.