PMDG 737: Air Systems panel is complete
The Air Systems panel is fully functional. However, we still need to add the ability to turn off the automatic announcements for the offsets for the panel. The automatic announcements for the left and right ram door indicators will be turned off by default. They turn on/off depending on whether the aircraft is airborn or on ground, and seem to turn on/off depending on ground/airspeed.