PMDG 737: MCP speed box
Replaced the current MCP speed box accessed with left bracket ([), SHIFT+S, with a new one. The replacement is fully functional except for the speedbrake lever. The detents do work though. To use the speedbrake, press ALT+K. When focused, press one of the following keys for the detents.

* O - off
* A - armed
* H - 50%
* F - Flt
* U - Up/100%

To use the autobrake, press ALT+B. When focused, use the following keys.

* R - RTO
* O - Off
* D - Disarm
* 1 - Autobrake 1
* 2 - Autobrake 2
* 3 - Autobrake 3

The replacement speed box also lets you change between mach and indicated air speed. To do so, either press ALT+C from anywhere in the speed box, or navigate to the C/O button and activate it.

TFM now alerts users when speed intervene is turned on/off. This is also indicated on the intervene button. To turn speed intervene on/off in the speed box, either press ALT+I to toggle it on/off, or navigate to the speed intervene button and activate it.

The N1 set selector is now included in the speed box. To use, either press ALT+L to cycle between the below options, or navigate to the N1 set selector button and activate it. Options include the following.

* 2 - Apply N1 to engine #2.
* 1 - Apply N1 to Engine #1.
* Both - Apply N1 equally to both engines.
Auto - Let the MCP handle N1.

Options 2, 1, and both are manual settings. In these instances, users will have to activate the N1 switch below.

The N1 switch is now included in the speed box. It will indicate its current state on the button and on the N1 indicator. To use, press ALT+N while in the speed box, or navigate to the N1 button and activate it.

Spoiler A/B are also included in the new speed box. To access, either press ALT+A for spoiler A, ALT+B for spoiler B, or navigate to the buttons and activate them.

The auto throttle is a carryover from the old speed box. It has the same behavior/actions as before. Either press ALT+U while in the speed box, or navigate to autothrottle and activate it.

The new speed box contains some indicators such as over/under speed, autothrottle, some speedbrake lights, and a few autobrake lights. Access them by either navigating to them with the TAB key, or pressing ALT+1 through ALT+9.