PMDG 737 MCP speed box +/- keys
Added the ability to increase/decrease the speedbrake with +/- (plus/minus) keys. The keys for jumping to the speedbrake detents also work. When using the +/- keys, the speedbrake value appears as a number between 101 and 399. The named detents will appear as words such as off, armed, half, flt, and up.