Change normal/on states to normal/off
Changed the normal/on states to normal/off to better represent a control's state with these available states.

GPWS flaps inhibit can be normal (on) or off, not normal (on) or on. Normal represents the normal working of the current control. Flaps inhibit in the normal position gives the impression that the flaps inhibit function is working normally. In the off position, the aircraft flaps warning function works normally.