SimBrief: Atis collections
The SimBrief flight plan leaves xml elements empty when no information about a section is not available, or the user chooses to leave it out when generating the flight plan. Added a check to ensure empty atis elements for the airports aren't crashing TFM. If there are empty atis elements, the AirportBlock.Atis list is null. Otherwise it has at least 1 AtisBlock inside it.