SimBrief: atis blocks
It appears that SimBrief changed the xml output in the API web service. It used to give more than 1 atis. Now, they are only giving real world atis for realism. TFM can handle more than one though. The trick is getting all atis objects for the airport in question. Once that is done, we can loop through the atis blocks and assign them to FlightPlan objects. We check for an existing list. If there is one, we wipe it out by setting it to null. We then continue making a new atis list and giving it atis objects. See the FlightPlan.LoadSimBriefOriginAsync method for an example.