SimBrief: waypoint more details window
Added a waypoint more details window. It shows the remaining properties of the waypoint selected in the navlog. It needs some cleanup and prettifying though. I need to remove the properties already listed in the navlog, remove the collection properties such as the wind data, round off some floating point values like mach, and change the title of the form to the waypoint name. It wouldn't hurt to remove the selection from the textbox to kill unwanted chatter from screen readers.