PMDG 737 FMC pages
The PMDG 737 FMC pages have button click events and key bindings assigned to them. All the key commands work so far. Sometimes NVDA will not read the current control name/state. The work around is to tab around to a few other controls to clear NVDA's buffer. Also added a new control that sits just to the right of the scratchpad. It is a single letter indicator (A or D) that indicates the current line select mode. A = alternate/F1-F12, and D = default (ctrl/alt 1-6). Access the indicator with ALT+A. The next new key command is CTRL+L, which is a convenience command that cycles between alternate and default line select modes. It will be functional in a later build.