PMDG737: Altitude window
Have most of the layout for the altitude window done. The toggle buttons use a byte to bool converter to change SingleStateToggle.CurrentState.Key to a true/false value. This should make things easier to show in control state (a toggle button pressed/not pressed). It is also easy to reverse the true/false value if needed. Works good with panel controls or indicators where the light is on to indicate something is turned off. An example of how to use it is in the BuildToggleButton method of the altitude window.