TFM: Styles dictionary
Added a Resources/Styles.xaml dictionary. The only thing in it is a style to turn a toggle button green with white text when it is switched on. Otherwise it keeps the default colors. To use it, put the following in the window where toggle buttons are found.

<Window.Resources>
        <ResourceDictionary>
            <ResourceDictionary.MergedDictionaries>
                <ResourceDictionary Source="/Resources/Styles.xaml"/>
            </ResourceDictionary.MergedDictionaries>
        </ResourceDictionary>
    </Window.Resources>

All of the toggle buttons will automatically take the new style. Also removed the value changed check from the toggle buttons because the behavior more reflects a toggle button. If the toggle button can't remain checked by the next task delay, the button returns to an unpressed state. Will change all of this to the other MCP boxes.