TFM: UI helpers
Added a new App partial in the Application folder called App.UIHelpers. It is a property of type UIHelpers and is named "UI". In the UIHelpers class, define any methods that help standardize the user interface. BuildToggleButton is the only one right now. Access it with App.UI.BuildToggleButton(...). Also moved the altitude box and the heading box over to the new App.UI helpers.