Treeview searching
Added a UIHelpers class that has methods for searching a treeview based on keywords. The method of interest is the SearchTreeView method. While in a WPF window with a treeview, Do the following:
* In the window's code behind, create a class level variable to hold the listview's original items.
List<TreeViewItem> originalItems;

* In the window's constructor, get the list of original items from your TreeView.
originalItems = App.UI.GetOriginalTreeViewItems(yourTreeView);
* In the event handler that performs your search, do the following.
App.UI.SearchTreeView(<add your parameters>);

The parameters are:
- Keyword: The comma separated list of search terms.
- treeview: The treeview to search.
- originalItems: The original treeview items from above.
- mappings: The dictionary<string, UserControlInfo> used to map keywords and user controls.

The search method does everything except focus related tasks. It does select the first treeview item if there is one. Also note that if there are no search results, the method inserts an item with the header 'No results found'. Also be aware of the UserControlInfo class required to make the search work. It is in the Properties folder. For an example on how to implement search in a tree, see the CockpitPanels in the PMDG737/CockpitPanels folder.