Move treeview items up and down
Added new methods to move items in a treeview up and down. The App.UI.MoveTreeViewItemUp and MoveTreeViewItemDown methods take a tree and a tree item as parameters. The tree is the target tree, and the tree item is the item to move. These methods also save the current tree state to disk so it can be reloaded later; for example, in a window's constructor. For a working example, see the cockpit panels dialog. To use the move up/down feature, press CTRL+UP or CTRL+DOWN while keyboard focus is on an item in the tree. You can move root items around (overhead, control stand...), and you can move child items within their panel. For example, you can move sections around in the overhead, but you can't move anything out of the overhead. The move up/down methods return true for a valid move and false for a failed move.