Sort treeviews in ascending/descending order
Added two methods to App.UI named SortTreeViewChildrenAscending and SortTreeViewChildrenDescending. Both thake the target treeview and the target treeview item. Both methods save the tree to desk after the sort. Also added a context menu to the PMDG 737 cockpit panels as an example of how to use the sort treeview methods.