Save/Load destination
Added a method to load the information for the destination runway window. When a pilot makes use of the destination runway window, they have the option of saving their destination information. If TFM ever were to crash or shutdown in the middle of a descent or landing, starting TFM again will load the previous entered information in the destination runway.

* While in the destination runway window, check the box for saving the destination.

Now, it will persist across restarts and reloading TFM will reactivate the destination.