Jump to runways
Fixed some visual aspects for the jump to runways dialog. The text field and its label are centered at the top of the dialog. The remaining space is taken by the runways list. NVDA repeats the cell's contents because it is reading the text located in the cell and its programmatic value at the same time. NVDA's navigator object will also present duplicate values. Unfortunately, there is nothing we can do about it. JAWS users should'nt have problems with duplicate announcements in data grids. Key commands do not work as of yet.