TFM connection and startup sequence
Streamlining the connection and startup timers.

* Remove unwanted code.
 - N1 limit for freeware aircraft.
 - Takeoff assist for freeware aircraft.
 - Timeout checks. TFM will now wait for a connection instead of timeout.

* Start moving IOSubsystem over to the Application class.
 - Each section of code has its own specialty. For example, utility code goes into the Application.Utilities class.
 - Non specialty code that most everything depends on goes directly into the Application class.
- Each specialty class or Application class segment should have its own file stored in the Application folder. The filename format is Application.{specialty}. EX: Application.UIHelpers or Application.ILSTracker.

So far, everything works as expected. I have redone the connection timer, the on startup event, and most of the high priority timer. There is also files and Application partials for UI helpers, keyboard events and handlers, ILS tracking, Utilities, and the main portion of Application found in App.xaml. There is still a large load of work to get done in this area, but a little each time someone is here will help.