Navigraph Header and Airport
Reconfigured the database context to accept the required connection string. Since the Navigraph database names tables with a tbl_ prefix, scaffolding also created the C# objects the same way. Starting with Header and Airport, we are going to reconfigure the database layer in TFM to make the database access code easier to understand. Scaffolding again will revert the code to its original form. However, we don't plan on doing a scaffold anytime soon. What small changes Navigraph makes to their database should not be difficult to change by hand.