Heartbeat conversion
We are starting to convert the heartbeat over to dotnet 7 and put it into the new WPF project structure. At this point, there are two instances of the heartbeat. Only one can run at a time. For now, it is the original heartbeat. However, having two copies of the heartbeat where the copy is broken, will not allow TFM to build. So, right now, TFM source builds are unusable. Expect build errors and thousands of compiletime errors. Use the official installer for production flights. We will try to get TFM up and running as soon as possible. Given the number of conversions/fixes required, source builds might be down for a while.